refactor(perf): replace KeyboardShortcutsProvider wrapper with sibling KeyboardShortcutsListener
Remove the "use client" KeyboardShortcutsProvider that wrapped the entire
app tree as a client boundary. Replace with KeyboardShortcutsListener, a
sibling component that handles keyboard events and the cheat sheet overlay
without wrapping children. Context sharing uses a module-level store with
useSyncExternalStore instead of React Context, eliminating the need for a
provider wrapper entirely.

Fixes #504
